+++ lib/config/configuration_test.go
@@ -1367,40 +1367,27 @@ func TestDatabaseFlags(t *testing.T) {
 func TestCheckAndSetDefaults(t *testing.T) {
 	tests := []struct {
 		comment      string
-		formatConfig string
+		formatConfig []string
 		hasError     bool
 	}{
 		{
 			comment:      "invalid key (does not exist)",
-			formatConfig: "{{level}} {{invalidkey}}",
-			hasError:     true,
-		},
-		{
-			comment:      "invalid log formatting, no curly braces with valid keys",
-			formatConfig: "level component",
-			hasError:     true,
-		},
-		{
-			comment:      "invalid log formatting, inconsistent curly braces with valid keys",
-			formatConfig: "{level}} {component}",
+			formatConfig: []string{"level", "invalid key"},
 			hasError:     true,
 		},
 		{
 			comment:      "valid keys and formatting",
-			formatConfig: "{{level}} {{component}} {{message}}",
-			hasError:     false,
-		},
-		{
-			comment:      "valid keys and formatting with inconsistent spacing",
-			formatConfig: "{{level}} {{message   }}",
+			formatConfig: []string{"level", "component", "message"},
 			hasError:     false,
 		},
 	}
 
-	formatter := &teleportFormatter{}
 	for _, tt := range tests {
 		t.Run(tt.comment, func(t *testing.T) {
-			err := formatter.CheckAndSetDefaults(tt.formatConfig)
+			formatter := &textFormatter{
+				InitFormat: tt.formatConfig,
+			}
+			err := formatter.CheckAndSetDefaults()
 			if tt.hasError {
 				require.Error(t, err)
 			} else {
