+++ lib/config/fileconf.go
@@ -460,7 +460,7 @@ type Log struct {
 	// Severity defines how verbose the log will be. Possible valus are "error", "info", "warn"
 	Severity string `yaml:"severity,omitempty"`
 	// Format defines the format in which each log line will be outputted. Example format: "{{timestamp}} {{component}} {{message}}"
-	Format string `yaml:"format,omitempty"`
+	Format []string `yaml:"format,omitempty"`
 }
 
 // Global is 'teleport' (global) section of the config file
